+++ RZBluetoothTests/RZBProfileBatteryTests.m
@@ -7,7 +7,7 @@
 //
 
 #import "RZBSimulatedTestCase.h"
-#import "CBPeripheral+RZBBattery.h"
+#import "RZBPeripheral+RZBBattery.h"
 #import "RZBSimulatedDevice+RZBBatteryLevel.h"
 #import "CBUUID+RZBPublic.h"
 #import "RZBLog.h"
