+++ RZBluetoothTests/RZBUserInteractionTests.m
@@ -7,7 +7,7 @@
 //
 
 #import "RZBSimulatedTestCase.h"
-#import "CBPeripheral+RZBBattery.h"
+#import "RZBPeripheral+RZBBattery.h"
 #import "RZBUserInteraction.h"
 #import "RZBSimulatedDevice+RZBBatteryLevel.h"
 #import "RZBErrors.h"
