+++ RZMockBluetooth/Simulation/RZBSimulatedDevice+RZBBatteryLevel.m
@@ -7,7 +7,7 @@
 //
 
 #import "RZBSimulatedDevice+RZBBatteryLevel.h"
-#import "CBPeripheral+RZBBattery.h"
+#import "RZBPeripheral+RZBBattery.h"
 #import "CBUUID+RZBPublic.h"
 
 static NSString *const RZBBatteryLevelKey = @"batteryLevel";
